Added utils.py in analytic sub-module
